put stats in own class
